# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/invoices": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "List available invoices",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "invoices"
                ],
                "summary": "List Invoices",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Status",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Invoice Number",
                        "name": "invoice_number",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Customer ID",
                        "name": "customer_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Created From",
                        "name": "created_from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Created To",
                        "name": "created_to",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Due From",
                        "name": "due_from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Due To",
                        "name": "due_to",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Finalized From",
                        "name": "finalized_from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Finalized To",
                        "name": "finalized_to",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Total Min",
                        "name": "total_min",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Total Max",
                        "name": "total_max",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Page Token",
                        "name": "page_token",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page Size",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/server.ListResponse"
                        }
                    }
                }
            }
        },
        "/invoices/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get invoice by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "invoices"
                ],
                "summary": "Get Invoice",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Invoice ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/server.DataResponse"
                        }
                    }
                }
            }
        },
        "/invoices/{id}/render": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Render invoice PDF/HTML",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "invoices"
                ],
                "summary": "Render Invoice",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Invoice ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.RenderInvoiceResponse"
                        }
                    }
                }
            }
        },
...
        "/price_amounts": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "List available price amounts",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "price_amounts"
                ],
                "summary": "List Price Amounts",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Price ID",
                        "name": "price_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Currency (ISO-4217)",
                        "name": "currency",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Effective From (RFC3339)",
                        "name": "effective_from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Effective To (RFC3339)",
                        "name": "effective_to",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Page Token",
                        "name": "page_token",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page Size",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/server.ListResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Create a new price amount",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "price_amounts"
                ],
                "summary": "Create Price Amount",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Idempotency Key",
                        "name": "Idempotency-Key",
                        "in": "header"
                    },
                    {
                        "description": "Create Price Amount Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_railzwaylabs_railzway_internal_priceamount_domain.CreateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/server.DataResponse"
                        }
                    }
                }
            }
        },
        "/price_amounts/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get price amount by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "price_amounts"
                ],
                "summary": "Get Price Amount",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Price Amount ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/server.DataResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "domain.RenderInvoiceResponse": {
            "type": "object",
            "properties": {
                "invoice_template_id": {
                    "type": "string"
                },
                "is_snapshot": {
                    "type": "boolean"
                },
                "rendered_html": {
                    "type": "string"
                },
                "rendered_pdf_url": {
                    "type": "string"
                }
            }
        },
...
        "github_com_railzwaylabs_railzway_internal_priceamount_domain.CreateRequest": {
            "type": "object",
            "properties": {
                "currency": {
                    "type": "string"
                },
                "effective_from": {
                    "type": "string"
                },
                "effective_to": {
                    "type": "string"
                },
                "maximum_amount_cents": {
                    "type": "integer"
                },
                "metadata": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "meter_id": {
                    "type": "string"
                },
                "minimum_amount_cents": {
                    "type": "integer"
                },
                "price_id": {
                    "type": "string"
                },
                "unit_amount_cents": {
                    "type": "integer"
                }
            }
        },
...
        "pagination.PageInfo": {
            "type": "object",
            "properties": {
                "has_more": {
                    "type": "boolean"
                },
                "next_page_token": {
                    "type": "string"
                },
                "previous_page_token": {
                    "type": "string"
                }
            }
        },
        "server.DataResponse": {
            "type": "object",
            "properties": {
                "data": {}
            }
        },
        "server.ListResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "page_info": {
                    "$ref": "#/definitions/pagination.PageInfo"
                }
            }
        },
...
    }
}